src/requirements.txt
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
boto3==1.34.19
requests==2.31.0
pytz==2023.3